+++ examples/basic-usage/commands/b/c/d.js
@@ -1,7 +1,5 @@
-module.exports = {
-  command: function (args, flags, context) {
-    console.log('d', args, flags, context)
-  },
-  args: [],
-  flags: []
+export function command(args, flags, context) {
+  console.log('b c d', args, flags, context)
 }
+export const args = []
+export const flags = []