+++ examples/basic-usage/commands/b/c/e.js
@@ -1,5 +1,5 @@
 export function command(args, flags, context) {
-  console.log('a', args, flags, context)
+  console.log('b c e', args, flags, context)
 }
 export const args = []
 export const flags = []