+++ examples/basic-usage/commands/b/c/index.js
@@ -1,7 +1,16 @@
-module.exports = {
-  command: function (args, flags, context) {
-    console.log('c', args, flags, context)
-  },
-  args: [],
-  flags: []
+export function command(args, flags, context) {
+  console.log('b c', args, flags, context)
+}
+export const args = []
+export const flags = []
+export const options = {
+  description: 'example description',
+  longDescription: 'this is an example of a nested command',
+  examples: [
+    {
+      command: 'example b c',
+      description: 'This executes the b c command'
+    }
+  ],
+  docsUrl: 'https://example.com'
 }