+++ examples/basic-usage/commands/e/f.js
@@ -1,5 +1,5 @@
 function command (args, flags, context) {
-
+  console.log('e f', args, flags, context)
 }
 
 const args = []
@@ -10,4 +10,4 @@ const options = {
   description: 'I am an f.'
 }
 
-module.exports = { command, args, flags, options }
+export default { command, args, flags, options }